requirements.txt
sounddevice==0.5.3
gpt4all==2.8.2
pyttsx3==2.99
faster-whisper==1.2.1
numpy==2.3.5
